+++ engine_store_ffi/src/ffi/sst_reader_impls.rs
@@ -5,10 +5,12 @@ use encryption::DataKeyManager;
 use engine_rocks::{get_env, RocksSstIterator, RocksSstReader};
 use engine_traits::{IterOptions, Iterator, RefIterable, SstReader};
 
-pub use super::interfaces::root::DB::{BaseBuffView, ColumnFamilyType};
 use super::{
-    interfaces::root::DB::SSTReaderInterfaces, LockCFFileReader, RaftStoreProxyPtr, RawVoidPtr,
-    SSTReaderPtr, SSTView, SSTViewVec,
+    interfaces::root::DB::{
+        BaseBuffView, ColumnFamilyType, RaftStoreProxyPtr, SSTReaderInterfaces, SSTReaderPtr,
+        SSTView, SSTViewVec,
+    },
+    LockCFFileReader, RawVoidPtr,
 };
 
 #[allow(clippy::clone_on_copy)]
